+++ CHANGES.md
@@ -7,6 +7,7 @@ Changed:
 Fixed:
 - Remote file resolution when passing URLs with spaces (#1410)
 - Fixed empty `{http,https}` body (#1417) 
+- Escape shell strings (#1215)
 
 1.4.3 (14-09-2020)
 =====